requirements.txt
openai>=1.3.0
sounddevice>=0.4.6
tiktoken>=0.5.1
faiss-cpu>=1.7.4
python-docx>=0.8.11
pdfminer.six>=20231228
PyMuPDF>=1.23.0
rich>=13.7.0
numpy>=1.24.0
scipy>=1.11.0
asyncio-mqtt>=0.16.0
toml>=0.10.2
pytest>=7.4.0 